requirements.txt
python-rpi.gpio
python3-rpi.gpio
python3-dev
python3-gpiozero
